refactor: split system tests across modules
This should make it easier to see which pandas tests
we have and which we are missing.
